==> notebooks/02.1-xlm.ipynb <==
%load_ext autoreload
%autoreload 2
import sys
sys.path.append("../src")

import mesh
from tqdm.autonotebook import tqdm

from transformers import XLMRobertaTokenizer, XLMRobertaModel, XLMRobertaConfig, XLMRobertaForMaskedLM

model_name = "xlm-roberta-base"
tokenizer = XLMRobertaTokenizer.from_pretrained(model_name)
model = XLMRobertaForMaskedLM.from_pretrained(model_name)

import torch

input_ids = tokenizer.encode("There is a dog by the window", add_special_tokens=True)

X = torch.tensor(input_ids).unsqueeze(0)

X

with torch.no_grad():
    outputs = model(X)    

preds = outputs[0].squeeze()

topk = preds.topk(10, axis=1)

topi = topk.indices.numpy()

for idx_list in topi.tolist():
    print(" ".join(tokenizer.decode(i) for i in idx_list))

tokenizer.decode([1,2,3,4,5])
