==> notebooks/03-babel-api.ipynb <==
%load_ext autoreload
%autoreload 2
import sys
sys.path.append("../src")

import mesh
from tqdm.autonotebook import tqdm
from CwnGraph import CwnBase
import json, pickle

import requests

with (mesh.get_data_dir()/"babelnet_key.txt").open("r") as fin:
    bn_key = fin.read().strip()
bn_url = "https://babelnet.io/v5/"

bn_key

bnapi = mesh.BabelNetAPI()

bnapi.get_version()

cwn = CwnBase()

lemmas = []
for nid, ndata in cwn.V.items():
    if ndata["node_type"] == "lemma":        
        lemmas.append(ndata["lemma"])

from itertools import chain
def find_cwn_senses(lemma):
    try:
        sense_iter = (x.senses for x in cwn.find_lemma(f"^{lemma}$"))
        sense_iter = chain.from_iterable(sense_iter)
        return list(sense_iter)
    except Exception as ex:
        print(lemma)
        print(ex)
        return []

target_lemmas = []
for lemma_x in tqdm(set(lemmas)):
    if not lemma_x:
        continue    
    try:        
        sense_list = find_cwn_senses(lemma_x)        
        if len(sense_list) > 2:
            target_lemmas.append(lemma_x)
    except Exception as ex:
        print(ex)        

target_lemma_path = mesh.get_data_dir()/"bn_target_lemmas.txt"
with target_lemma_path.open("w", encoding="UTF-8") as fout:
    for lemma in target_lemmas:
        fout.write(lemma + "\n")

target_lemma_path = mesh.get_data_dir()/"bn_target_lemmas.txt"
with target_lemma_path.open("r", encoding="UTF-8") as fin:
    taget_lemmas = fin.readlines()

## Continue to retrieve data from BabelNet

target_lemma_path = mesh.get_data_dir()/"bn_target_lemmas.txt"
with target_lemma_path.open("r", encoding="UTF-8") as fin:
    target_lemmas = [x.strip() for x in fin.readlines()]
    
bn_sense_path = mesh.get_data_dir()/"bn_sense_data.pkl"
if bn_sense_path.exists():
    with bn_sense_path.open("rb") as fin:
        sense_data = pickle.load(fin)
else:
    sense_data = {}

len(sense_data)

target_lemmas[0]

quota = 900
for lemma_x in tqdm(target_lemmas):
    if quota == 0:
        break
    if lemma_x in sense_data:
        continue
    data = bnapi.get_senses(lemma_x)
    if 'message' in data:        
        print(data["message"])
        break
    quota -= 1
    sense_data[lemma_x] = data    

len(sense_data)

with bn_sense_path.open("wb") as fout:
    pickle.dump(sense_data, fout)

s1 = list(sense_data.items())[1]

to_deleted = []
for lemma, senses in sense_data.items():
    if 'message' in senses:
        to_deleted.append(lemma)

for lemma in to_deleted:
    sense_data.pop(lemma)

len(sense_data)

wn_map = {}
for lemma, senses in sense_data.items():    
    if not senses:
        continue
    
    wn_senses = filter(lambda x: x["type"] == "WordNetSense", senses)    
    slist = []
    for s in wn_senses:
        
        p = s["properties"]
        slist.append((p.get("wordNetOffset"), 
                     p.get("fullLemma"), 
                     p.get("senseKey"), p.get("pos"), p.get("synsetID").get("id"), p.get("bKeySense")))
    wn_map[lemma] = slist

print(f"target_lemmas (>3 in CWN): {len(target_lemmas)}\n"
      f"sense_data (requested from BN so far): {len(sense_data)}\n"
      f"with WN senses: {len(wn_map)}")

## A word sense sample

from nltk.corpus import wordnet as wn

wn_map["犯"]

wn.synset_from_pos_and_offset("v", 2582615).definition()

wn.synset_from_pos_and_offset("v", 2582921).definition()

find_cwn_senses("犯")

## Make a dataframe

import pandas as pd

data = [(lemma, *map_entry) for lemma, sense_map in wn_map.items() for map_entry in sense_map]
    

sense_map_frame = pd.DataFrame.from_records(data)

sense_map_frame

sense_dir = mesh.get_data_dir()/"sense_data"
mesh.ensure_dir(sense_dir)
sense_map_frame.to_csv(sense_dir/"sense_map_bn_pwn.csv")
