==> notebooks/03.1-cwn-map.ipynb <==
%load_ext autoreload
%autoreload 2
import sys
sys.path.append("../src")

import mesh
import pandas as pd

sense_dir = mesh.get_data_dir()/"sense_data"
mesh.ensure_dir(sense_dir)
sense_map_frame = pd.read_csv(sense_dir/"sense_map_bn_pwn.csv", index_col=0)

from CwnGraph import CwnBase
cwn = CwnBase()
cwn.find_lemma("^聊$")[2].senses

TARGETS = "^[說談講聊]$"
sub_frame = sense_map_frame.loc[sense_map_frame.word.str.match(TARGETS), :]

sub_frame = sub_frame.groupby("pwn_synset").first().sort_values("word").reset_index()

from nltk.corpus import wordnet as wn
def get_definition(syn_id):
    syn_pos = syn_id[-1]
    syn_num = syn_id[:-1]
    return wn.synset_from_pos_and_offset(syn_pos, int(syn_num)).definition()

def get_examples(syn_id):
    syn_pos = syn_id[-1]
    syn_num = syn_id[:-1]
    exs = wn.synset_from_pos_and_offset(syn_pos, int(syn_num)).examples()
    return "\n".join(exs)

sub_frame = sub_frame.assign(
    definition=lambda df: [get_definition(x) for x in df.pwn_synset],
    examples=lambda df: [get_examples(x) for x in df.pwn_synset]
)

sub_frame.index.name="serial"
sub_frame.to_csv(sense_dir/"pwn_definitions_about_talk.csv", encoding="UTF-8")

# CWN Senses

from itertools import chain
def find_cwn_senses(lemma):
    try:
        sense_iter = (x.senses for x in cwn.find_lemma(f"^{lemma}$"))
        sense_iter = chain.from_iterable(sense_iter)
        return list(sense_iter)
    except Exception as ex:
        print(lemma)
        print(ex)
        return []

data = []
for w in "說談講":
    senses = find_cwn_senses(w)
    for sense_x in senses:
        data.append(dict(
            word = w,
            sense_pos = sense_x.pos,
            sense_id = sense_x.id,
            sense_def = sense_x.definition,
            sense_examples = "\n".join(sense_x.all_examples())
        ))

cwn_frame = pd.DataFrame.from_records(data)

cwn_frame.index.name = "serial"

cwn_frame.head()

cwn_frame.to_csv(sense_dir/"cwn_senses_about_talk.csv", encoding="UTF-8")
