==> notebooks/03.2-cwn-pwn-ontoAlign.ipynb <==
# OntoAlignment between PWN and CWN

* What is the data structure necessary to align ontologically between PWN and CWN

%load_ext autoreload
%autoreload 2
import sys
sys.path.append("../src")

import mesh
import pandas as pd
from tqdm.autonotebook import tqdm

sense_dir = mesh.get_data_dir()/"sense_data"
mesh.ensure_dir(sense_dir)
sense_map_frame = pd.read_csv(sense_dir/"sense_map_bn_pwn.csv", index_col=0)

entry = sense_map_frame.iloc[8, :].to_dict()

entry

from CwnGraph import CwnBase
cwn = CwnBase()

from nltk.corpus import wordnet as wn
def get_synset(syn_id):
    syn_pos = syn_id[-1]
    syn_num = syn_id[:-1]
    return wn.synset_from_pos_and_offset(syn_pos, int(syn_num))

cwn_senses = cwn.find_all_senses(entry["word"])

syn = get_synset(entry["pwn_synset"])

mesh.get_pwn_relations(entry["pwn_synset"], depth=2)

wn.synsets("cat")[0].name()

dir(wn.synsets("cat")[0])
mesh.get_pwn_relations(wn.synsets("cat")[0], ["hypernyms"], depth=2)

from mesh import cwn_onto_align as onto_align

adata = onto_align.get_alignment_structure(entry["word"], entry["pwn_synset"], cwn)

align_data = []
for idx, row in tqdm(sense_map_frame.iterrows(), total=sense_map_frame.shape[0]):
    adata = onto_align.get_alignment_structure(row["word"], row["pwn_synset"], cwn)
    align_data.append(adata)

## Generate mapping json

import json
with open(mesh.get_data_dir()/"sense_data/bn_alignment_data.json", "w", encoding="UTF-8") as fout:
    json.dump([x.to_dict() for x in align_data], fout, indent=2, ensure_ascii=False)

import pickle
with open(mesh.get_data_dir()/"sense_data/bn_alignment_data.pkl", "wb") as fout:
    pickle.dump(align_data, fout)

import nltk
nltk.download('omw')

## CWN existing PWN alignment

niter = filter(lambda x: x[1]['node_type']==("pwn_synset"), cwn.V.items())

synsets=list(niter)

synsets[0]

cwn.find_edges(synsets[1][0], is_directed=False)

mesh.get_synset("04355268n")

wn.synsets("endowment")[0].lemmas()

from CwnGraph import CwnSense
CwnSense("07057003", cwn)

## Read PWN-1.6

wn16_dir = mesh.get_data_dir()/"wordnet-1.6"
noun_path = wn16_dir / "dict/data.noun"
verb_path = wn16_dir / "dict/data.verb"
adj_path = wn16_dir / "dict/data.adj"
adv_path = wn16_dir / "dict/data.adv"

def parse_synset(ln_txt):
    syn_data, gloss = ln_txt.split("|")    
    toks = syn_data.split()    
    offset = toks[0]
    pos = toks[2]
    n_lemma = int(toks[3], 16)    
    lemmas = toks[slice(4,n_lemma*2+4,2)]
    return (offset, pos, lemmas, gloss.strip())

parse_synset(noun_data[15226])

toks = "00001740 03 n 02 entity 0 something 0 014".split()

def load_wn16_data(data_path, map_data):
    with data_path.open() as fin:
        data = [x for x in fin.readlines() if not x.startswith(" ")]
    for ln in tqdm(data, desc=data_path.name):
        syn_data = parse_synset(ln)
        map_data[syn_data[0]+syn_data[1]] = syn_data    

map_data = {}
load_wn16_data(noun_path, map_data)
load_wn16_data(verb_path, map_data)
load_wn16_data(adj_path, map_data)
load_wn16_data(adv_path, map_data)

len(map_data)

sense_dir = mesh.get_data_dir() / "sense_data"
wn16_synpath = sense_dir/"wn16_synsets.json"
with wn16_synpath.open("w", encoding="UTF-8") as fout:
    json.dump(map_data, fout, ensure_ascii=False, indent=2)

## Load PWN synset data

from CwnGraph import PwnSynset, CwnBase
import json

cwn = CwnBase()

niter = filter(lambda x: x[1]['node_type']==("pwn_synset"), cwn.V.items())

pwn_synsets=list(niter)

len(pwn_synsets)

cwn.find_edges(pwn_synsets[0][0], False)

syn=PwnSynset("pwn_09553033N", cwn)

syn.relations

### load WN16 data

sense_dir = mesh.get_data_dir() / "sense_data"
wn16_synpath = sense_dir/"wn16_synsets.json"
with wn16_synpath.open("r", encoding="UTF-8") as fin:
    wn16_data = json.load(fin)

wn16_data["09553033n"]

entries = []
for syn in tqdm(pwn_synsets):
    syn_id, syn_data = syn
    syn_offset = syn_id.split("_")[1].lower()
    
    wn16_syn = None
    if syn_offset in wn16_data:
        wn16_syn = wn16_data[syn_offset]
    
    if syn_offset.replace("a", "s") in wn16_data:
        wn16_syn = wn16_data[syn_offset.replace("a", "s")]
    
    if not wn16_syn:
        print("Cannot find " + syn_offset)
        continue
        
    pwn_synset = PwnSynset(syn_id, cwn)
    rels = pwn_synset.relations     
    for rel_x in rels:
        rel_type, cwn_node, _ = rel_x
        try:
            entries.append(dict(
                cwn_id=cwn_node.id,
                cwn_pos=cwn_node.pos,
                lemmas=cwn_node.lemmas[0].lemma,
                cwn_def=cwn_node.definition,
                rel_type=rel_type,
                wn16_offset=wn16_syn[0],
                wn16_pos=wn16_syn[1],
                wn16_lemmas=','.join(wn16_syn[2]),
                wn16_def=wn16_syn[3]
            ))            
        except Exception as ex:
            print(syn_id, end=" ")
            print(ex)                 

import pandas as pd

wn16_map = pd.DataFrame.from_records(entries)

wn16_map.to_csv(sense_dir / "wn16_cwn_map.csv", encoding="UTF-8")
