==> notebooks/03.3-cwn-pwn-16-30-map.ipynb <==
# Mapping between PWN 1.6 and 3.0

%load_ext autoreload
%autoreload 2
import sys
sys.path.append("../src")

import mesh
import pandas as pd
from tqdm.autonotebook import tqdm

from nltk.corpus import wordnet as wn

wn.get_version()

## load PWN1.6 data

import json

import numpy as np

sense_dir = mesh.get_data_dir() / "sense_data"
wn16_map = pd.read_csv(sense_dir / "wn16_cwn_map.csv", encoding="UTF-8", index_col=0, 
                       dtype={"cwn_id": np.object, "wn16_offset": np.object})

wn16_uniq = wn16_map.groupby("wn16_offset").first().reset_index()

from fuzzywuzzy import fuzz
import re

def map_to_wn30(wn16_entry):
    wn16_lemmas = wn16_entry.wn16_lemmas.split(",")
    wn16_def = wn16_entry.wn16_def.split(";")[0]    
    head = re.sub(r"\(\w\)", "", wn16_lemmas[0])
    wn30_syns = wn.synsets(head)
    if not wn30_syns:
        raise ValueError("Cannot find %s in WN30" % (head,))
    candid_vec = []
    
    for syn_x in wn30_syns:
        wn30_lemmas = set(syn_x.lemmas())
        score = fuzz.ratio(wn16_def, syn_x.definition())
        candid_vec.append((syn_x, score))
    candid_vec.sort(key=lambda x: x[1], reverse=True)
    if not candid_vec:
        breakpoint()
    return candid_vec[0][0], candid_vec[0][1]

map_data = []
for row_id, entry in tqdm(wn16_uniq.iterrows(),total=wn16_uniq.shape[0]):    
    try:
        wn30_syn, score = map_to_wn30(entry)
        map_data.append(dict(
            cwn_id=entry.cwn_id,
            wn16_def=entry.wn16_def,
            wn16_lemmas=entry.wn16_lemmas,
            wn16_offset=entry.wn16_offset,
            wn16_pos=entry.wn16_pos,
            wn30_syn_name=wn30_syn.name(),
            wn30_lemmas=",".join([x.name() for x in wn30_syn.lemmas()]),
            wn30_def=wn30_syn.definition(),
            score=score
        ))
    except Exception as ex:        
        map_data.append(dict(
            cwn_id=entry.cwn_id,
            wn16_def=entry.wn16_def,
            wn16_lemmas=entry.wn16_lemmas,
            wn16_offset=entry.wn16_offset,
            wn30_syn_name=None,
            wn30_lemmas=None,
            wn30_def=None,
            score=-1,
        ))


cwn_wn16_wn30_map = pd.DataFrame.from_records(map_data)

cwn_wn16_wn30_map

cwn_wn30_map_checked = cwn_wn16_wn30_map.loc[cwn_wn16_wn30_map.score>=62, :]

cwn_wn16_wn30_map.to_csv(sense_dir/"cwn_wn16_wn30_map.csv", encoding="UTF-8")

cwn_wn30_map_checked.to_csv(sense_dir/"cwn_wn16_wn30_map.checked.csv", encoding="UTF-8")

print(cwn_wn16_wn30_map.shape, cwn_wn30_map_checked.shape)
