==> notebooks/03.4-cwn-wn30.ipynb <==
%load_ext autoreload
%autoreload 2
import sys
sys.path.append("../src")

import mesh
import pandas as pd
import numpy as np
from tqdm.autonotebook import tqdm

from nltk.corpus import wordnet as wn

sense_dir = mesh.get_data_dir() / "sense_data"
wn30_map = pd.read_csv(sense_dir / "cwn_wn16_wn30_map.checked.csv", encoding="UTF-8", index_col=0,
                      dtype={"cwn_id": np.object, "wn16_offset": np.object})

wn30_map["wn16_key"] = wn30_map.wn16_offset + wn30_map.wn16_pos
wn30_map = wn30_map.set_index("wn16_key")

## Update CWN PwnSynset

from CwnGraph import CwnBase

cwn = CwnBase()

niter = filter(lambda x: x[1]['node_type']==("pwn_synset"), cwn.V.items())

cwn.V[pwn_syn[0]]

def find_wn30_map(pwn_id):
    wn16_key = pwn_id.split("_")[1].lower()
    buf = [wn16_key, wn16_key.replace("a", "s")]
    while buf:
        key = buf.pop(0)        
        if key in wn30_map.index:
            return wn30_map.loc[key].to_dict()
    return None    

mapped_count = 0
niter = filter(lambda x: x[1]['node_type']==("pwn_synset"), cwn.V.items())
for pwn_id, ndata in tqdm(niter):
    wn30_data = find_wn30_map(pwn_id)
    if not wn30_data:
        continue
    mapped_count += 1
    ndata.update(dict(
        wn30_name=wn30_data["wn30_syn_name"],
        wn30_lemmas=wn30_data["wn30_lemmas"],
        wn30_def=wn30_data["wn30_def"]        
    ))    
    
print(mapped_count)

cwn.V[pwn_syn[0]]

## Write CwnGraph data structure

import pickle
cwn_data_path = sense_dir/"cwn_graph_0.2.pkl"
with cwn_data_path.open("wb") as fout:
    cwn.meta = {"version": "0.2", "description": "add WN30 map to pwn_synset node (build 001)"}
    pickle.dump((cwn.V, cwn.E, cwn.meta), fout)

## Install new data as CWN default 
CwnBase.install_cwn(cwn_data_path)

cwn = CwnBase()

cwn.V[pwn_syn[0]]

cwn.find_edges(pwn_syn[0])
