==> notebooks/03.5-cwn-wn30-pwnSynset.ipynb <==
%load_ext autoreload
%autoreload 2

from CwnGraph import CwnBase, CwnSense

cwn = CwnBase()

cwn.find_all_senses("資")[2].pwn_synsets

pwn_syn.hypernym_paths()

sense_iter = filter(lambda x: x[1].get("node_type", "") == "sense", cwn.V.items())

senses = list(sense_iter)

len(senses)

senses_list = map(lambda x: CwnSense(x[0], cwn), senses)
senses_list = filter(lambda x: x.pwn_synsets and any(y[1].synset_wn30_name for y in x.pwn_synsets), senses_list)
senses_list = list(senses_list)

len(senses_list)

pwn_list = []
for sense_x in senses_list:
    try:
        head = sense_x.lemmas[0].lemma
    except:
        head = "----"
    pwn_list.append(dict(
        cwn_id = sense_x.id,
        lemma = head,
        definition = sense_x.definition, 
        wn30 = [(x[0], str(x[1])) for x in sense_x.pwn_synsets if x[1].synset_wn30_name]
    ))

len(pwn_list)

pwn_list[10]

sense_dir = mesh.get_data_dir() / "sense_data"

import json
import sys
sys.path.append("../src")
import mesh

sense_dir = mesh.get_data_dir() / "sense_data"
cwn_senses_wn30_path = sense_dir / "cwn_senses_wn30.json"
with cwn_senses_wn30_path.open("w", encoding="UTF-8")  as fout:
    json.dump(pwn_list, fout, indent=2, ensure_ascii=False)
