==> notebooks/03.6-pwn-core.ipynb <==
from mesh_import import mesh

wn30_core_path = mesh.get_data_dir() / "sense_data/wn30-core-synsets.tab"

wn30_cores = [x.strip() for x in wn30_core_path.open().readlines()]

from nltk.corpus import wordnet as wn
import pandas as pd

syn_offset, syn_pos = wn30_cores[0].split("-")
syn = wn.synset_from_pos_and_offset(syn_pos, int(syn_offset))

data = []
for syn_id in wn30_cores:
    syn_offset, syn_pos = syn_id.split("-")
    syn = wn.synset_from_pos_and_offset(syn_pos, int(syn_offset))
    data.append(dict(
        syn_id=syn_offset,
        syn_pos=syn.pos(),
        syn_name=syn.name(),
        lemmas=",".join([x.name() for x in syn.lemmas()]),
        definition=syn.definition()
    ))
core_frame = pd.DataFrame().from_records(data)
core_frame.syn_id.astype(np.object)
core_frame.dtypes

core_frame

import numpy as np
core_frame.to_csv(mesh.get_data_dir() / "sense_data/core_synset_wn30.csv")
