==> notebooks/04.1-beh-data.ipynb <==
%load_ext autoreload
%autoreload 2
import sys
sys.path.append("../src")

import mesh
from tqdm.autonotebook import tqdm

import pandas as pd
from CwnGraph import CwnBase, CwnLemma

meld_sch=pd.read_excel(mesh.get_data_dir()/"ldt-data/Tsang-2018-MELD-SCH.xlsx")

clpw = pd.read_excel(mesh.get_data_dir()/"ldt-data/Tse-2017-Chinese-lexicon-project.xlsx", sheet_name=0)

cwn = CwnBase()

# get lemmas with at least one sense
lemma_iter = filter(lambda x: x[1]['node_type']=='lemma', cwn.V.items())
lemma_iter = filter(lambda x: CwnLemma(x[0], cwn).senses, lemma_iter)
lemma_iter = map(lambda x: x[1].get("lemma"), lemma_iter)

cwn_lemmas = list(lemma_iter)

## CLD Dataset

cld = pd.read_csv(mesh.get_data_dir()/"CLD/chineselexicaldatabase2.1.txt")

from opencc import OpenCC
s2t = OpenCC('s2t')
cld=cld.assign(trad_word=lambda df: [s2t.convert(x) for x in df.Word])

sub_cld = cld[["trad_word", "FrequencyRaw"]]

## Graph measures

cns = pd.read_csv(mesh.get_data_dir()/"cns/cwn_centralities.csv", index_col=0)

cns.head()

## MELD_SCH & CWN coverage

cn_words = meld_sch.loc[meld_sch.lexicality==1,].word.tolist()

from opencc import OpenCC
cc = OpenCC('s2t')

tw_words = [cc.convert(x) for x in cn_words]

int_words = set(cwn_lemmas) & set(tw_words)
print(
    f"MELD_SCH Words: {len(cn_words)}\n"
    f"Intersect with CWN: {len(int_words)}")

meld_sch = meld_sch.assign(trad_word=lambda df: [s2t.convert(x) for x in df.word])

meld_sch_words = pd.DataFrame(data={"word":list(int_words)})

meld_sch.head()

meld_sch_frame = meld_sch_words.merge(meld_sch.loc[meld_sch.lexicality==1,:][["trad_word", "N", "RT", "RTSD", "ERR", "length"]], 
                                    left_on="word", right_on="trad_word")
meld_sch_frame = meld_sch_frame.merge(cns)
# meld_sch_frame = meld_sch_frame.merge(sub_cld, left_on="word", right_on="trad_word")

meld_sch_words.shape

meld_sch_frame.tail()

meld_sch_frame.to_csv(mesh.get_data_dir()/"cns/cns_meld_sch.csv", encoding="UTF-8")

## CLP & CWN Coverage

clp_words = clpw.Word_Trad.tolist()

int_words = set(cwn_lemmas) & set(clp_words)
print(
    f"CLP Words: {len(clp_words)}\n"
    f"Intersect with CWN: {len(int_words)}")
